(refactor) Move change password modal out of user profile
